+++ src/decomp/decomp.c
@@ -2,7 +2,7 @@
 
 #include "game/background.h"
 #include "game/camera.h"
-#include "game/console.h"
+#include "game/console/common.h"
 #include "game/effects.h"
 #include "game/game.h"
 #include "game/gameflow.h"
