+++ src/game/console/cmd/die.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <libtrx/game/console/common.h>
+
+extern CONSOLE_COMMAND g_Console_Cmd_Die;
