+++ src/game/console/common.c
@@ -1,4 +1,4 @@
-#include "game/console.h"
+#include "game/console/common.h"
 
 #include "game/clock.h"
 #include "game/game_string.h"
